Bug #705
I have implemented URL encoding for letter-characters. Previously, all characters that were not A-Z, 0-9, "-", or "_" were removed. Now the code is a little more intelligent and checks the unicode category of each character. If it is a "letter", it will be kept. This includes accented characters, and probably characters from other alphabets too. These characters will be URL encoded, so they show as %C3%B1, for example.

Interestingly, Firefox 3 will show "/monta%C3%B1a" as "montaña" in the address bar, which is nice.

I have updated both the client-side Javascript, and everywhere on the server-side that creates a URL from a string, to use the new formatting (it's implemented in the ToSafeUrl() string extension method).